requirements.txt
python-dotenv
livekit-agents[openai, silero, turn-detector, elevenlabs, groq]
livekit-plugins-noise-cancellation
PyYAML
qdrant-client
aiohttp
httpio
sentence-transformers
textual
pyinstaller
pytz

#API
fastapi
uvicorn
pydantic